+++ internal/lsptrace/embed.go
@@ -1,0 +1,6 @@
+package lsptrace
+
+import _ "embed"
+
+//go:embed ui.html
+var uiHTML []byte